bug fixed

+++ server/action.py
@@ -37,16 +37,19 @@ def fill_shared_memory(shared_memory, data):
 
 def free_communication_system(shared_memory_name, path_tube_1, path_tube_2):
     try:
-        os.unlink(path_tube_1)
+        if(os.path.exists("/tmp/tubenommeprincipalsecond.fifo")):
+            os.unlink(path_tube_1)
     except OSError as error:
         print("Warning : ", error)
     try:
-        os.unlink(path_tube_2)
+        if(os.path.exists("/tmp/tubenommesecondprincipal.fifo")):
+            os.unlink(path_tube_2)
     except OSError as error:
         print("Warning : ", error)
     try:
-        shared_memory_to_delete = shared_memory.SharedMemory(shared_memory_name)
-        shared_memory_to_delete.unlink()
+        if(os.path.exists("/run/shm/leclerc")):
+            shared_memory_to_delete = shared_memory.SharedMemory(shared_memory_name)
+            shared_memory_to_delete.unlink()
     except OSError as error:
         print("Warning : ", error)
